refactor: Move all exercises into their own file, add jsdoc for exported constructs
